refactor: move bucket to new structure
