+++ cmd/influx/ping.go
@@ -13,8 +13,9 @@ func newPingCmd() *cli.Command {
 		Before: middleware.WithBeforeFns(withCli(), withApi(false)),
 		Flags:  coreFlags,
 		Action: func(ctx *cli.Context) error {
+			cli := getCLI(ctx)
 			client := ping.Client{
-				CLI: getCLI(ctx),
+				CLI: &cli,
 				API: getAPINoToken(ctx).HealthApi,
 			}
 			return client.Ping(ctx.Context)
